smartgas-debit/backend/requirements.txt
Flask==3.0.3
flask-apispec==0.11.4
marshmallow==3.22.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
requests==2.32.3
pytz==2024.2
Flask-Cors==5.0.0
sentry-sdk[flask]==2.17.0
